+++ Platform/AptioInputFix/Pointer/AIM.h
@@ -11,14 +11,21 @@ THE PROGRAM IS DISTRIBUTED UNDER THE BSD LICENSE ON AN "AS IS" BASIS,
 WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 
 **/
-#ifndef _AMI_SHIM_POINTER_H_
-#define _AMI_SHIM_POINTER_H_
+
+#ifndef AIM_SELF_H
+#define AIM_SELF_H
+
+#include <Library/UefiLib.h>
+#include <Protocol/AmiPointer.h>
+#include <Protocol/SimplePointer.h>
 
 // Taken from AMI
-#define MAX_POINTERS 6
-#define POSITION_POLL_TIMER_INTERVAL 50000
-#define AMI_SHIM_TIMER_PERIOD 50000
-#define SCALE_FACTOR 1 // 1~4
+#define AIM_MAX_POINTERS 6
+// AMI has 66666 here, but we need to sync with AppleEvent.
+// Slightly lower resolution results in sometimes overflowng mouse, but
+// it does not miss the keys in response.
+#define AIM_POSITION_POLL_INTERVAL EFI_TIMER_PERIOD_MILLISECONDS(10)
+#define AIM_SCALE_FACTOR 1 // 1~4
 
 typedef struct {
   EFI_HANDLE                    DeviceHandle;
@@ -35,10 +42,18 @@ typedef struct {
   BOOLEAN                       TimersInitialised;
   BOOLEAN                       UsageStarted;
   EFI_EVENT                     ProtocolArriveEvent;
-  UINTN                         OriginalTimerPeriod;
-  EFI_TIMER_ARCH_PROTOCOL       *TimerProtocol;
   EFI_EVENT                     PositionEvent;
-  AMI_SHIM_POINTER_INSTANCE     PointerMap[MAX_POINTERS];
+  AMI_SHIM_POINTER_INSTANCE     PointerMap[AIM_MAX_POINTERS];
 } AMI_SHIM_POINTER;
+
+EFI_STATUS
+AIMInit (
+  VOID
+  );
+
+EFI_STATUS
+AIMExit (
+  VOID
+  );
 
 #endif