+++ Changelog.md
@@ -2,6 +2,7 @@ AptioFix Changelog
 ==================
 
 #### R24
+- Added basic APTIO V compatibility to AptioInputFix
 - Fixed broken timer resolution in AptioInputFix
 - Disabled ALT/CMD swapping in AptioInputFix